+++ mars/metrics/tests/test_metric_api.py
@@ -13,9 +13,16 @@
 # limitations under the License.
 
 import pytest
+from random import random
 
 from .. import api
-from ..api import init_metrics, Metrics
+from ..api import (
+    init_metrics,
+    Metrics,
+    _percentile_builder,
+    Percentile,
+    record_time_cost_percentile,
+)
 
 
 @pytest.fixture
@@ -86,3 +93,48 @@ def test_histogram():
     for i in range(2002):
         h.record(1)
     assert h.value > 0
+
+
+def test_percentile_report():
+    def gen_callback(data):
+        def callback(value):
+            data.append(value)
+
+        return callback
+
+    data90 = []
+    data95 = []
+    data99 = []
+
+    all_data = []
+    percentile_args = [
+        (Percentile.PercentileType.P90, gen_callback(data90), 100),
+        (Percentile.PercentileType.P95, gen_callback(data95), 100),
+        (Percentile.PercentileType.P99, gen_callback(data99), 100),
+    ]
+    percentile_list = [
+        _percentile_builder[percentile_type](callback, window)
+        for percentile_type, callback, window in percentile_args
+    ]
+    for _ in range(199):
+        data = random()
+        all_data.append(data)
+        for percentile in percentile_list:
+            percentile.record_data(data)
+    sub_data = sorted(all_data[:100])
+    print(sub_data[:10])
+    assert len(data90) == 1 and sub_data[10 - 1] == data90[0]
+    assert len(data95) == 1 and sub_data[5 - 1] == data95[0]
+    assert len(data99) == 1 and sub_data[1 - 1] == data99[0]
+
+
+def test_invaild_percentile_report():
+    with pytest.raises(ValueError):
+        Percentile(-1, 10, lambda x: ...)
+
+    with pytest.raises(ValueError):
+        Percentile(1, -1, lambda x: ...)
+
+    with pytest.raises(ValueError):
+        with record_time_cost_percentile([]):
+            raise ValueError
