+++ mars/tests/test_utils.py
@@ -27,7 +27,6 @@ from concurrent.futures import ThreadPoolExecutor
 from functools import partial
 from io import BytesIO
 from enum import Enum
-from random import random
 
 import numpy as np
 
@@ -42,7 +41,6 @@ from .. import tensor as mt
 from .. import utils
 from ..core import tile, TileableGraph
 from ..serialization.ray import register_ray_serializers
-from ..metrics.api import _percentile_builder, Percentile, record_time_cost_percentile
 from .core import require_ray
 
 
@@ -588,48 +586,3 @@ def test_get_func_token_values():
 
     func = Func()
     assert get_func_token_values(func) == [func]
-
-
-def test_percentile_report():
-    def gen_callback(data):
-        def callback(value):
-            data.append(value)
-
-        return callback
-
-    data90 = []
-    data95 = []
-    data99 = []
-
-    all_data = []
-    percentile_args = [
-        (Percentile.PercentileType.P90, gen_callback(data90), 100),
-        (Percentile.PercentileType.P95, gen_callback(data95), 100),
-        (Percentile.PercentileType.P99, gen_callback(data99), 100),
-    ]
-    percentile_list = [
-        _percentile_builder[percentile_type](callback, window)
-        for percentile_type, callback, window in percentile_args
-    ]
-    for _ in range(199):
-        data = random()
-        all_data.append(data)
-        for percentile in percentile_list:
-            percentile.record_data(data)
-    sub_data = sorted(all_data[:100])
-    print(sub_data[:10])
-    assert len(data90) == 1 and sub_data[10 - 1] == data90[0]
-    assert len(data95) == 1 and sub_data[5 - 1] == data95[0]
-    assert len(data99) == 1 and sub_data[1 - 1] == data99[0]
-
-
-def test_invaild_percentile_report():
-    with pytest.raises(ValueError):
-        Percentile(-1, 10, lambda x: ...)
-
-    with pytest.raises(ValueError):
-        Percentile(1, -1, lambda x: ...)
-
-    with pytest.raises(ValueError):
-        with record_time_cost_percentile([]):
-            raise ValueError
